+++ src/postgres/query/constraints/key_column_usage.rs
@@ -1,8 +1,10 @@
 #[derive(Debug, sea_query::Iden)]
 /// Ref: https://www.postgresql.org/docs/13/infoschema-key-column-usage.html
 pub enum KeyColumnUsageFields {
-    ContraintSchema,
+    ConstraintCatalog,
+    ConstraintSchema,
     ConstraintName,
+    TableCatalog,
     TableSchema,
     TableName,
     ColumnName,