+++ tests/migration/src/lib.rs
@@ -1,0 +1,16 @@
+use sea_schema::migration::prelude::*;
+
+mod m20220118_000001_create_cake_table;
+mod m20220118_000002_create_fruit_table;
+
+pub struct Migrator;
+
+#[async_trait::async_trait]
+impl MigratorTrait for Migrator {
+    fn migrations() -> Vec<Box<dyn MigrationTrait>> {
+        vec![
+            Box::new(m20220118_000001_create_cake_table::Migration),
+            Box::new(m20220118_000002_create_fruit_table::Migration),
+        ]
+    }
+}